webapp/requirements.txt
-e git://github.com/philomat/django-persistent-messages.git#egg=django-persistent-messages
django-postmark==0.1.6
